Moved new data utilities into their own package and left a stub deprecated Spreadsheet class behind to be removed in the future.
