+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/DataImporter.java
@@ -27,8 +27,8 @@ import com.adobe.acs.commons.mcp.form.FileUploadComponent;
 import com.adobe.acs.commons.mcp.form.FormField;
 import com.adobe.acs.commons.mcp.form.RadioComponent;
 import com.adobe.acs.commons.mcp.model.GenericReport;
-import com.adobe.acs.commons.mcp.util.CompositeVariant;
-import com.adobe.acs.commons.mcp.util.Spreadsheet;
+import com.adobe.acs.commons.data.CompositeVariant;
+import com.adobe.acs.commons.data.Spreadsheet;
 import java.io.IOException;
 import java.lang.reflect.Array;
 import java.util.ArrayList;
