+++ bundle/src/main/java/com/adobe/acs/commons/mcp/impl/processes/asset/UrlAssetImport.java
@@ -26,8 +26,8 @@ import com.adobe.acs.commons.functions.CheckedConsumer;
 import com.adobe.acs.commons.mcp.ProcessInstance;
 import com.adobe.acs.commons.mcp.form.FileUploadComponent;
 import com.adobe.acs.commons.mcp.form.FormField;
-import com.adobe.acs.commons.mcp.util.Spreadsheet;
-import com.adobe.acs.commons.mcp.util.CompositeVariant;
+import com.adobe.acs.commons.data.Spreadsheet;
+import com.adobe.acs.commons.data.CompositeVariant;
 import com.day.cq.commons.jcr.JcrUtil;
 import com.day.cq.dam.api.Asset;
 import java.io.IOException;
