+++ bundle/src/test/java/com/adobe/acs/commons/data/VariantTest.java
@@ -17,8 +17,10 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.mcp.util;
+package com.adobe.acs.commons.data;
 
+import com.adobe.acs.commons.data.CompositeVariant;
+import com.adobe.acs.commons.data.Variant;
 import java.text.SimpleDateFormat;
 import java.time.Instant;
 import java.util.Date;
