+++ bundle/src/test/java/com/adobe/acs/commons/mcp/util/Spreadsheet.java
@@ -1,0 +1,49 @@
+/*
+ * #%L
+ * ACS AEM Commons Bundle
+ * %%
+ * Copyright (C) 2018 Adobe
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package com.adobe.acs.commons.mcp.util;
+
+import java.io.IOException;
+import java.io.InputStream;
+import org.apache.sling.api.request.RequestParameter;
+
+/**
+ * @Deprecated Class was moved to com.adobe.acs.commons.data.Spreadsheet
+ */
+public class Spreadsheet extends com.adobe.acs.commons.data.Spreadsheet {    
+    public Spreadsheet(boolean convertHeaderNames, String... headerArray) {
+        super(convertHeaderNames, headerArray);
+    }
+
+    public Spreadsheet(boolean convertHeaderNames, InputStream file, String... required) throws IOException {
+        super(convertHeaderNames, file, required);
+    }
+
+    public Spreadsheet(boolean convertHeaderNames, RequestParameter file, String... required) throws IOException {
+        super(convertHeaderNames, file, required);
+    }
+
+    public Spreadsheet(InputStream file, String... required) throws IOException {
+        super(file, required);
+    }
+
+    public Spreadsheet(RequestParameter file, String... required) throws IOException {
+        super(file, required);
+    }    
+}
